signed short stereo audio on mac

+++ include/tiny3d.h
@@ -24,14 +24,14 @@
 	#endif
 	void open_window(int width, int height); //width = 0: fullscreen
 	//define this:
-	extern void update(double time, double deltaTime, int width, int height, int nSamples, float *samples);
+	extern void update(double time, double deltaTime, int width, int height, int nSamples, signed short *samples);
 #else
 	#define SCREEN_WIDTH 256
 	#define SCREEN_HEIGHT 192
 	extern uint32_t screen[SCREEN_WIDTH*SCREEN_HEIGHT];
 	void open_window(int scale); //scale = 0: fullscreen
 	//define this:
-	extern void update(double time, double deltaTime, int nSamples, float *samples);
+	extern void update(double time, double deltaTime, int nSamples, signed short *samples);
 
 	//drawing functions:
 	void clear_screen(uint32_t color);
